roots/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.51.0",
    "mcp[cli]>=1.8.0",
    "openai==2.30.0",
    "prompt-toolkit>=3.0.51",
    "pyboxen>=1.3.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]
